requirements.txt
fastapi
uvicorn[standard]
pymongo
passlib[bcrypt]
python-jose
PyJWT
pydantic[email]
python-dotenv